+++ pybamm/models/full_battery_models/lead_acid/base_lead_acid_model.py
@@ -82,3 +82,12 @@ class BaseModel(pybamm.BaseBatteryModel):
 
         self.submodels["negative sei"] = pybamm.sei.NoSEI(self.param, "Negative")
         self.submodels["positive sei"] = pybamm.sei.NoSEI(self.param, "Positive")
+
+    def set_lithium_plating_submodel(self):
+
+        self.submodels["negative lithium plating"] = pybamm.lithium_plating.NoPlating(
+            self.param, "Negative"
+        )
+        self.submodels["positive lithium plating"] = pybamm.lithium_plating.NoPlating(
+            self.param, "Positive"
+        )