+++ tests/unit/test_parameters/test_parameter_values.py
@@ -93,6 +93,12 @@ class TestParameterValues(unittest.TestCase):
             param.update({"b": 1})
 
     def test_check_parameter_values(self):
+        # Cell capacity [A.h] deprecated
+        with self.assertRaisesRegex(ValueError, "Cell capacity"):
+            pybamm.ParameterValues({"Cell capacity [A.h]": 1,
+                                    "Nominal cell capacity [A.h]": 1})
+        with self.assertWarnsRegex(DeprecationWarning, "Cell capacity"):
+            pybamm.ParameterValues({"Cell capacity [A.h]": 1})
         # Can't provide a current density of 0, as this will cause a ZeroDivision error
         with self.assertRaisesRegex(ValueError, "Typical current"):
             pybamm.ParameterValues({"Typical current [A]": 0})
@@ -324,6 +330,13 @@ class TestParameterValues(unittest.TestCase):
         processed_diff_func = parameter_values.process_symbol(diff_func)
         self.assertEqual(processed_diff_func.evaluate(inputs={"a": 3}), 123)
 
+        # make sure diff works, despite simplifications, when the child is constant
+        a_const = pybamm.Scalar(3)
+        func_const = pybamm.FunctionParameter("func", {"a": a_const})
+        diff_func_const = func_const.diff(a_const)
+        processed_diff_func_const = parameter_values.process_symbol(diff_func_const)
+        self.assertEqual(processed_diff_func_const.evaluate(), 123)
+
         # function parameter that returns a python float
         func = pybamm.FunctionParameter("float_func", {"a": a})
         processed_func = parameter_values.process_symbol(func)
@@ -375,16 +388,16 @@ class TestParameterValues(unittest.TestCase):
 
         parameter_values = pybamm.ParameterValues({"Diffusivity": D})
 
-        a = pybamm.InputParameter("a")
+        a = pybamm.Scalar(3)
         func = pybamm.FunctionParameter("Diffusivity", {"a": a})
 
         processed_func = parameter_values.process_symbol(func)
-        self.assertEqual(processed_func.evaluate(inputs={"a": 3}), 9)
+        self.assertEqual(processed_func.evaluate(), 9)
 
         # process differentiated function parameter
         diff_func = func.diff(a)
         processed_diff_func = parameter_values.process_symbol(diff_func)
-        self.assertEqual(processed_diff_func.evaluate(inputs={"a": 3}), 6)
+        self.assertEqual(processed_diff_func.evaluate(), 6)
 
     def test_multi_var_function_with_parameters(self):
         def D(a, b):
@@ -460,15 +473,15 @@ class TestParameterValues(unittest.TestCase):
             check_already_exists=False,
         )
 
-        a = pybamm.InputParameter("a")
+        a = pybamm.Scalar(0.6)
         func = pybamm.FunctionParameter("function", {"a": a})
         interp = pybamm.FunctionParameter("interpolation", {"a": a})
 
         processed_func = parameter_values.process_symbol(func)
         processed_interp = parameter_values.process_symbol(interp)
         np.testing.assert_array_almost_equal(
-            processed_func.evaluate(inputs={"a": 0.6}),
-            processed_interp.evaluate(inputs={"a": 0.6}),
+            processed_func.evaluate(),
+            processed_interp.evaluate(),
             decimal=4,
         )
 
@@ -478,10 +491,17 @@ class TestParameterValues(unittest.TestCase):
         processed_diff_func = parameter_values.process_symbol(diff_func)
         processed_diff_interp = parameter_values.process_symbol(diff_interp)
         np.testing.assert_array_almost_equal(
-            processed_diff_func.evaluate(inputs={"a": 0.6}),
-            processed_diff_interp.evaluate(inputs={"a": 0.6}),
+            processed_diff_func.evaluate(),
+            processed_diff_interp.evaluate(),
             decimal=2,
         )
+
+    def test_process_not_constant(self):
+        param = pybamm.ParameterValues({"a": 4})
+
+        a = pybamm.NotConstant(pybamm.Parameter("a"))
+        self.assertIsInstance(param.process_symbol(a), pybamm.NotConstant)
+        self.assertEqual(param.process_symbol(a).evaluate(), 4)
 
     def test_process_complex_expression(self):
         var1 = pybamm.Variable("var1")